pyproject.toml
[tool.poetry]
name = "listener"
version = "0.1.0"
description = ""
authors = ["Sergei Solodskikh <ssolodskikh@adapty.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyaudio = "^0.2.14"
speechrecognition = "^3.10.1"
pydub = "^0.25.1"
openai = "^1.13.3"
whisper = "^1.1.10"
keyboard = "^0.13.5"
python-dotenv = "^1.0.1"
pyperclip = "^1.8.2"
pyautogui = "^0.9.54"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
